# swagger.yml
swagger: "2.0"
info:
  title: Subscription Aggregation API
  description: API for managing and aggregating user subscription data
  version: "1.0.0"
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

paths:
  /subscriptions:
    post:
      tags:
        - Subscriptions
      summary: Create a new subscription
      description: Creates a new subscription record
      parameters:
        - in: body
          name: subscription
          description: Subscription data
          required: true
          schema:
            $ref: "#/definitions/SubscriptionCreate"
      responses:
        201:
          description: Subscription created successfully
          schema:
            $ref: "#/definitions/Subscription"
        400:
          description: Invalid input data
        500:
          description: Internal server error

    get:
      tags:
        - Subscriptions
      summary: List all subscriptions
      description: Returns a list of subscriptions with optional filtering
      parameters:
        - in: query
          name: user_id
          type: string
          format: uuid
          description: Filter by user ID
        - in: query
          name: service_name
          type: string
          description: Filter by service name
        - in: query
          name: offset
          type: integer
          description: Number of items for pagination offset
        - in: query
          name: limit
          type: integer
          description: Number of items per page. Default 20, hard limit 1000
        - in: query
          name: from
          type: string
          format: date
          pattern: "MM-YYYY"
          description: Filter by Start date that later or equal specified date in MM-YYYY format
        - in: query
          name: to
          type: string
          format: date
          pattern: "MM-YYYY"
          description: Filter by Start date that earlier or equal specified date in MM-YYYY format. Cannot be earlier that from param. If from and to are equal then filter by specific date only.
      responses:
        200:
          description: A list of subscriptions
          schema:
            type: array
            items:
              $ref: "#/definitions/Subscription"
        400:
          description: Invalid filter data
        500:
          description: Internal server error

  /subscription/{id}:
    get:
      tags:
        - Subscriptions
      summary: Get a subscription by ID
      description: Returns a single subscription
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: ID of the subscription to get
      responses:
        200:
          description: Subscription found
          schema:
            $ref: "#/definitions/Subscription"
        404:
          description: Subscription not found
        500:
          description: Internal server error

    put:
      tags:
        - Subscriptions
      summary: Update a subscription
      description: Updates an existing subscription
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: ID of the subscription to update
        - in: body
          name: subscription
          description: Updated subscription data
          required: true
          schema:
            $ref: "#/definitions/SubscriptionUpdate"
      responses:
        200:
          description: Subscription updated successfully
          schema:
            $ref: "#/definitions/Subscription"
        400:
          description: Invalid input data
        404:
          description: Subscription not found
        500:
          description: Internal server error

    delete:
      tags:
        - Subscriptions
      summary: Delete a subscription
      description: Deletes a subscription record
      parameters:
        - in: path
          name: id
          type: string
          format: uuid
          required: true
          description: ID of the subscription to delete
      responses:
        204:
          description: Subscription deleted successfully
        404:
          description: Subscription not found
        500:
          description: Internal server error

  /subscriptions/sum:
    get:
      tags:
        - Summary
      summary: Sum subscription costs
      description: Calculates total cost of subscriptions for a given period with optional filters
      parameters:
        - in: query
          name: from
          type: string
          format: date
          pattern: "MM-YYYY"
          required: true
          description: Start date in MM-YYYY format
        - in: query
          name: to
          type: string
          format: date
          pattern: "MM-YYYY"
          required: true
          description: End date in MM-YYYY format. Cannot be earlier that Start date (from)
        - in: query
          name: user_id
          type: string
          format: uuid
          description: Filter by user ID
        - in: query
          name: service_name
          type: string
          description: Filter by service name
      responses:
        200:
          description: Summary result
          schema:
            $ref: "#/definitions/SummaryResult"
        400:
          description: Invalid date format or parameters
        500:
          description: Internal server error

definitions:
  SubscriptionCreate:
    type: object
    required:
      - service_name
      - price
      - user_id
      - start_date
    properties:
      service_name:
        type: string
        example: "Yandex Plus"
        description: Name of the subscription service
      price:
        type: integer
        example: 400
        description: Monthly subscription cost in rubles (without kopecks)
      user_id:
        type: string
        format: uuid
        example: "60601fee-2bf1-4721-ae6f-7636e79a0cba"
        description: User ID in UUID format
      start_date:
        type: string
        format: date
        pattern: "MM-YYYY"
        example: "07-2025"
        description: Start date of subscription in MM-YYYY format
      end_date:
        type: string
        format: date
        pattern: "MM-YYYY"
        example: "12-2025"
        description: Optional end date of subscription in MM-YYYY format

  SubscriptionUpdate:
    type: object
    properties:
      service_name:
        type: string
        example: "Yandex Plus Premium"
        description: Name of the subscription service
      price:
        type: integer
        example: 500
        description: Monthly subscription cost in rubles (without kopecks)
      start_date:
        type: string
        format: date
        pattern: "MM-YYYY"
        example: "08-2025"
        description: Start date of subscription in MM-YYYY format
      end_date:
        type: string
        format: date
        pattern: "MM-YYYY"
        example: "12-2025"
        description: Optional end date of subscription in MM-YYYY format
        nullable: true

  Subscription:
    type: object
    properties:
      id:
        type: string
        format: uuid
        example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        description: Subscription ID
      service_name:
        type: string
        example: "Yandex Plus"
      price:
        type: integer
        example: 400
      user_id:
        type: string
        format: uuid
        example: "60601fee-2bf1-4721-ae6f-7636e79a0cba"
      start_date:
        type: string
        format: date
        pattern: "MM-YYYY"
        example: "07-2025"
      end_date:
        type: string
        format: date
        pattern: "MM-YYYY"
        example: "12-2025"
        nullable: true

  SummaryResult:
    type: object
    properties:
      sum:
        type: integer
        example: 1200
        description: Total cost of all matching subscriptions for the period